WTA #44: Renamed classes.
